backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg[binary]==3.2.12

